+++ plugs/directive/complete.ts
@@ -10,7 +10,7 @@ export async function queryComplete(completeEvent: CompleteEvent) {
   const allEvents = await events.listEvents();
 
   return {
-    from: completeEvent.pos - match[2].length,
+    from: completeEvent.pos - match[1].length,
     options: allEvents
       .filter((eventName) => eventName.startsWith("query:"))
       .map((source) => ({