+++ web/components/page_navigator.tsx
@@ -1,13 +1,18 @@
 import { FilterList } from "./filter.tsx";
 import { FilterOption, PageMeta } from "../../common/types.ts";
+import { CompletionContext, CompletionResult } from "../deps.ts";
 
 export function PageNavigator({
   allPages,
   onNavigate,
+  completer,
+  vimMode,
   currentPage,
 }: {
   allPages: Set<PageMeta>;
+  vimMode: boolean;
   onNavigate: (page: string | undefined) => void;
+  completer: (context: CompletionContext) => Promise<CompletionResult | null>;
   currentPage?: string;
 }) {
   const options: FilterOption[] = [];
@@ -40,7 +45,8 @@ export function PageNavigator({
       placeholder="Page"
       label="Open"
       options={options}
-      // icon={faFileLines}
+      vimMode={vimMode}
+      completer={completer}
       allowNew={true}
       helpText="Start typing the page name to filter results, press <code>Return</code> to open."
       newHint="Create page"